# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is API documentation for Project.",
        "title": "Qpay API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/gateway": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all gateways.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all gateways",
                "responses": {
                    "200": {
                        "description": "List of gateways",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.GatewayResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting gateways\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/admin/gateway/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific gateway by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get gateway by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gateway details",
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayResponse"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Gateway not found!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the status of a specific gateway by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update gateway status by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated status details",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.StatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated gateway details",
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Status field is unsupported\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
        "/gateway": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all gateways associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "List all gateways for the authenticated user",
                "responses": {
                    "200": {
                        "description": "List of gateways",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.GatewayResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"You didn't add any gateway. Please register a gateway!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register a new gateway with the provided details.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Register a new gateway for the authenticated user",
                "parameters": [
                    {
                        "description": "Gateway details",
                        "name": "gatewayRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Bind Error\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "403": {
                        "description": "{\"status\": \"error\", \"message\": \"Commission is incorrect\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in gateway\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/gateway/commission/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a list of all active commissions.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "List all active commissions",
                "responses": {
                    "200": {
                        "description": "List of commissions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.CommissionResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "{\"status\": \"error\", \"message\": \"Internal server error in getting commissions\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/gateway/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve details of a specific gateway associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Find a gateway by ID for the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gateway details",
                        "schema": {
                            "$ref": "#/definitions/handlers.GatewayResponse"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Gateway ID is not correct\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "404": {
                        "description": "{\"status\": \"error\", \"message\": \"Gateway does not exist!\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update details of a specific gateway associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Update a gateway for the authenticated user",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated gateway details",
                        "name": "updateRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateGatewayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Invalid gateway details\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
        "/gateway/{id}/address": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Purchase an address for a specific gateway associated with the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateways"
                ],
                "summary": "Purchase an address for a gateway",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Gateway ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Purchase address details",
                        "name": "purchaseAddressRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PurchaseAddressRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success message",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    },
                    "400": {
                        "description": "{\"status\": \"error\", \"message\": \"Address already in use\"}",
                        "schema": {
                            "$ref": "#/definitions/handlers.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BankAccountResponse": {
            "type": "object",
            "properties": {
                "account_owner": {
                    "type": "string"
                },
                "bank_logo": {
                    "type": "string"
                },
                "bank_name": {
                    "type": "string"
                },
                "sheba": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.CommissionResponse": {
            "type": "object",
            "properties": {
                "Percent_per_transaction": {
                    "type": "number"
                },
                "amount_per_transaction": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.GatewayRequest": {
            "type": "object",
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "commission_id": {
                    "type": "integer"
                },
                "is_personal": {
                    "type": "boolean"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "handlers.GatewayResponse": {
            "type": "object",
            "properties": {
                "bank_account": {
                    "$ref": "#/definitions/handlers.BankAccountResponse"
                },
                "commission": {
                    "$ref": "#/definitions/handlers.CommissionResponse"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "route": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.PurchaseAddressRequest": {
            "type": "object",
            "properties": {
                "route": {
                    "type": "string"
                }
            }
        },
...
        "handlers.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "handlers.StatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateGatewayRequest": {
            "type": "object",
            "properties": {
                "bank_account_id": {
                    "type": "integer"
                },
                "commission_id": {
                    "type": "integer"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}